The CITES database package

citesdb provides over 40 years of wildlife trade data from the Convention on International Trade in Endangered Species of Wild Fauna and Flora. These data on imports and exports of CITES annex species are provided by CITES member countries to the CITES secretariat.

Details

This package allows you to download the full database of historic transactions (commonly called the CITES Trade Database) to analyze locally on your computer. Because of the large size of this data, the citesdb package stores it as an on-disk MonetDB that can be queried without loading into RAM.

See also